修改价格的类型为float

+++ create.sql
@@ -11,7 +11,7 @@ CREATE TABLE `product` (
   `link` text COLLATE utf8_unicode_ci COMMENT '产品链接',
   `shop_name` varchar(255) COLLATE utf8_unicode_ci DEFAULT NULL COMMENT '店铺名',
   `sales_num` int(10) unsigned DEFAULT NULL COMMENT '销售数量',
-  `price` decimal(12,2) DEFAULT '0.00' COMMENT '价格',
+  `price` FLOAT DEFAULT '0.00' COMMENT '价格',
   PRIMARY KEY (`id`),
   UNIQUE KEY (`name`)
 ) ENGINE=InnoDB AUTO_INCREMENT=1 DEFAULT CHARSET=utf8 COLLATE=utf8_unicode_ci;
@@ -20,9 +20,9 @@ CREATE TABLE `product` (
 DROP TABLE IF EXISTS `product_info`;
 CREATE TABLE `product_info` (
   `id` int(11) NOT NULL AUTO_INCREMENT COMMENT '主键，自增',
-   `name` varchar(255) COLLATE utf8_unicode_ci DEFAULT NULL COMMENT '产品名',
-  `origin_price` decimal(12,2) DEFAULT NULL COMMENT '原价',
-  `real_price` decimal(12,2) DEFAULT NULL COMMENT '实际卖价',
+  `name` varchar(255) COLLATE utf8_unicode_ci DEFAULT NULL COMMENT '产品名',
+  `origin_price` FLOAT DEFAULT NULL COMMENT '原价',
+  `real_price` FLOAT DEFAULT NULL COMMENT '实际卖价',
   `brand` varchar(255) COLLATE utf8_unicode_ci DEFAULT NULL COMMENT '品牌名',
   `address` varchar(64) COLLATE utf8_unicode_ci DEFAULT NULL COMMENT '产地,省份城市',
   `category` varchar(64) COLLATE utf8_unicode_ci DEFAULT NULL COMMENT '菊花茶种类',
